% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCols.R
\name{addCols}
\alias{addCols}
\title{Internal Function used copy the model dataframe and add additional columns}
\usage{
addCols(fittedModel)
}
\arguments{
\item{fittedModel}{is a logistic regression model fitted using the glm function}
}
\description{
Internal Function used copy the model dataframe and add additional columns
}
\examples{
new_data <- addCols(fittedModel)

}
\keyword{fitted}
