% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGLM.R
\name{coef.compareGLM}
\alias{coef.compareGLM}
\title{Logistic Regression Coefficients and Confidence Intervals}
\usage{
\method{coef}{compareGLM}(x, exp = F, raw = F, ci = 0.95, ci_normal = F, sigfig = 6, expand = F)
}
\arguments{
\item{x}{An object of class 'compareGLM' i.e. a list of fitted \code{glm} model object(s) of family \code{binomial}.}

\item{exp}{\code{logical}. If \code{TRUE} then exponentiate the coefficients and confidence intervals; if \code{FALSE} do not exponentiate (default).}

\item{raw}{\code{logical}. If \code{TRUE} then return the values in a \code{data.frame} object, if \code{FALSE} return the values in a \code{gt} table object (default).}

\item{ci}{The confidence interval level required.}

\item{ci_normal}{\code{logical}. If \code{TRUE} then calculate the confidence interval based on asymptotic normality; if \code{FALSE} calculate the profile likelihood confidence interval using \code{confint} (default).}

\item{sigfig}{The number of significant figures to round the results to.}

\item{expand}{\code{logical}. If \code{TRUE} then additionally return the standard errors, z values and p-values from the glm summary object.}
}
\description{
S3 method for class 'compareGLM' used to present the model coefficient estimates and confidence intervals of one or more \code{glm} objects in a gt table or data.frame.
}
\examples{
modelList <- compareGLM(fittedModel1, fittedModel2)
coef(modelList, exp=T)

}
\keyword{coefficients,}
\keyword{confidence}
\keyword{forest}
\keyword{interval,}
\keyword{plot}
