% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGLM.R
\name{coef.listGLM}
\alias{coef.listGLM}
\title{Logistic Regression Coefficients and Confidence Intervals}
\usage{
\method{coef}{listGLM}(x, exp = F, raw = F, ci = 0.95, ci_normal = F, sigfig = 6, expand = F)
}
\arguments{
\item{x}{An object of class 'listGLM' i.e. a list of fitted \code{glm} model object(s) of family \code{binomial}.}

\item{exp}{\code{logical}. If \code{TRUE} then exponentiate the coefficients and confidence intervals; if \code{FALSE} do not exponentiate (default).}

\item{raw}{\code{logical}. If \code{TRUE} then return the values in a \code{list} of \code{data.frame} objects, if \code{FALSE} return the values in a \code{gt} group object (default).}

\item{ci}{The confidence interval level required. Default is 0.95.}

\item{ci_normal}{\code{logical}. If \code{TRUE} then calculate the confidence interval based on asymptotic normality; if \code{FALSE} calculate the profile likelihood confidence interval using \code{confint} (default).}

\item{sigfig}{The number of significant figures to round the results to. Default is 6.}

\item{expand}{\code{logical}. If \code{TRUE} then additionally return the z values and p-values from the glm summary object.}
}
\description{
S3 method for class 'listGLM' used to present the model coefficient estimates, standard errors and confidence intervals of one or more \code{glm} objects in a gt table or data.frame.
}
\examples{
modelList <- listGLM(fittedModel1, fittedModel2)
coef(modelList, exp=T)

}
\keyword{coefficients,}
\keyword{confidence}
\keyword{forest}
\keyword{interval,}
\keyword{plot}
