% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.listGLM.R
\name{compare}
\alias{compare}
\title{Compare GLM models using various criteria}
\usage{
compare(x, raw = F, sigfig = 4, expand = F)
}
\arguments{
\item{x}{An object of class 'listGLM' i.e. a list of fitted \code{glm} model object(s) of family \code{binomial}.}

\item{raw}{\code{logical}. If \code{TRUE} then return the values in a \code{data.frame} object, if \code{FALSE} return the values in a \code{gt} table object (default).}

\item{sigfig}{The number of significant figures to round the results to. Default is 4.}

\item{expand}{\code{logical}. If\code{FALSE} (default), return the AIC, BIC and McFadden's Pseudo R2-values. If \code{TRUE} then return additional Pseudo-R2 values.}
}
\description{
\code{compare} is generic function which calculates the AIC, BIC and Pseudo-R-squared values for each model in a list.
}
\examples{
fit_glm1 <- glm(am ~ cyl + hp + wt, data=mtcars, family = binomial(link = "logit"))
fit_glm2 <- glm(am ~ cyl + hp, data=mtcars, family = binomial(link = "logit"))
modelList <- listGLM(fit_glm1, fit_glm2)
compare(modelList, raw=TRUE, expand=TRUE)

}
\keyword{R-squared}
\keyword{aic,}
\keyword{bic,}
\keyword{compare,}
\keyword{pseudo}
