% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGLM.R
\name{compareGLM}
\alias{compareGLM}
\title{Compare GLM models}
\usage{
compareGLM(...)
}
\arguments{
\item{...}{one or more fitted \code{glm} model objects of family \code{binomial}.}
}
\description{
Function used create list of glm objects of class 'compareGLM'
}
\examples{
x <- compareGLM(fittedModel1, fittedModel2, fittedModel3)


}
\keyword{glm}
\keyword{list}
