% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGLM.R
\name{getCoef}
\alias{getCoef}
\title{Get coefficients from GLM models}
\usage{
getCoef(fittedModels, exp = F, ci = 0.95, ci_normal = F, sigfig = 6)
}
\arguments{
\item{fittedModels}{A list of fitted \code{glm} model objects of family \code{binomial}, of class 'listGLM'.}

\item{exp}{\code{logical}. If \code{TRUE} then exponentiate the coefficients and confidence intervals; if \code{FALSE} do not exponentiate (default).}

\item{ci}{The confidence interval level required.}

\item{ci_normal}{\code{logical}. If \code{TRUE} then calculate the confidence interval based on asymptotic normality; if \code{FALSE} calculate the profile likelihood confidence interval using \code{confint} (default).}

\item{sigfig}{The number of significant figures to round the results to.}
}
\description{
Get coefficients and other values from GLM models, used internally by methods for the 'listGLM' class.
}
\examples{
coef <- getCoef(listGLM(fittedModel1, fittedModel2, fittedModel3))

}
\keyword{coefficients,}
\keyword{confidence}
\keyword{glm}
\keyword{interval,}
