% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Internal rcpp function - Likelihood Ratio Test}
\alias{Internal rcpp function - Likelihood Ratio Test}
\alias{rcpp_multipleLRTest}
\title{Internal rcpp function - Likelihood Ratio Test}
\usage{
rcpp_multipleLRTest(dev_vec, df_vec, comparison_matrix)
}
\description{
Internal rcpp function - Likelihood Ratio Test
}
\details{
Internal rcpp function that performs a likelihood ratio test between pairs of models. Returns the p-values and model indicies.
}
\examples{
list_LR <-multipleLRTest(m_dev, m_df, comparison_matrix)
}
