% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGLM.R
\name{listGLM}
\alias{listGLM}
\title{List GLM models}
\usage{
listGLM(...)
}
\arguments{
\item{...}{one or more fitted \code{glm} model objects of family \code{binomial}.}
}
\description{
Function used create list of glm objects of class 'listGLM'
}
\examples{
x <- listGLM(fittedModel1, fittedModel2, fittedModel3)


}
\keyword{glm}
\keyword{list}
