% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleLRTest.R
\name{multipleLRTest}
\alias{multipleLRTest}
\title{Likelihood Ratio Test}
\usage{
multipleLRTest(model_list, alpha = 0.05, override = F)
}
\arguments{
\item{model_list}{A list of fitted \code{glm} models object of family \code{binomial}.}

\item{alpha}{The significance level of the LR test. Default is 0.05.}

\item{override}{\code{logical}. If \code{TRUE} then perform the LR test for all pairs in the input list; if \code{FALSE} then only perform the LR test for nested model pairs.}
}
\description{
Performs a likelihood ratio test between pairs of models in a given list.
}
\examples{
new_data <- addCols(fittedModel)

}
\keyword{fitted}
