% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGLM.R
\name{plot.compareGLM}
\alias{plot.compareGLM}
\title{Forest Plot of Logistic Regression Coefficient Estimates}
\usage{
\method{plot}{compareGLM}(x, exp = F, ci = 0.95, ci_normal = F, sigfig = 6)
}
\arguments{
\item{x}{An object of class 'compareGLM' i.e. a list of fitted \code{glm} model object(s) of family \code{binomial}.}

\item{exp}{\code{logical}. If \code{TRUE} then the coefficients are exponentiated, so the x marker is 0; if \code{FALSE}, x marker is 1.}
}
\description{
S3 method for class 'compareGLM' used to plot coefficient estimates and confidence intervals in a Forest Plot.
}
\examples{
modelList <- compareGLM(fittedModel1, fittedModel2)
plot(modelList)

}
\keyword{confidence}
\keyword{forest}
\keyword{interval}
\keyword{plot,}
