% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCoef.R
\name{plotCoef}
\alias{plotCoef}
\title{Forest Plot of Logistic Regression Coefficient Estimates}
\usage{
plotCoef(coef_data, exp = F)
}
\arguments{
\item{coef_data}{A \code{data.frame} containing the coefficient estimates and confidence intervals of one or more \code{glm} models.}

\item{exp}{\code{logical}. If \code{TRUE} then the coefficients are exponentiated, so the x marker=0; if \code{FALSE}, x marker=1.}
}
\description{
Internal Method used to plot coefficient estimates and confidence intervals in a Forest Plot. Used by \code{viewCoef}.
}
\examples{
plotCoef(fittedModel)


}
\keyword{confidence}
\keyword{forest}
\keyword{interval}
\keyword{plot,}
