% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredictors.R
\name{plotPredictors.glm}
\alias{plotPredictors.glm}
\title{Plot Logit of Response against Numerical Predictors}
\usage{
\method{plotPredictors}{glm}(fittedModel, smooth = T, interactive = F)
}
\arguments{
\item{fittedModel}{A fitted \code{glm} model object of family \code{binomial}.}

\item{smooth}{\code{logical}. If \code{TRUE} then include a loess smoothing line (default); if \code{FALSE} do not include the loess line.}

\item{interactive}{\code{logical}. If \code{TRUE} then return an interactive plot; if \code{FALSE}, return ggplot objects (default).}
}
\description{
S3 Method for glm object used to return plots of the logit of the response vs numerical predictors. Used during model diagnostics.
}
\examples{
fit_glm1 <- glm(am ~ cyl + hp + wt, data=mtcars, family = binomial(link = "logit"))
my_plots <- plotPredictors(fit_glm1)
my_plots[[1]]

}
\keyword{assumptions,}
\keyword{diagnostics,}
\keyword{linearity,}
\keyword{logit,}
\keyword{plots}
\keyword{predictor}
