% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewCoef.R
\name{viewCoef}
\alias{viewCoef}
\title{Logistic Regression Coefficients and Confidence Intervals}
\usage{
viewCoef(
  fittedModels,
  exp = F,
  plot = F,
  raw = F,
  ci = 0.95,
  ci_normal = F,
  sigfig = 6,
  expand = F
)
}
\arguments{
\item{fittedModels}{A single or list of fitted \code{glm} model object of family \code{binomial}.}

\item{exp}{\code{logical}. If \code{TRUE} then exponentiate the coefficients and confidence intervals; if \code{FALSE} do not exponentiate (default).}

\item{plot}{\code{logical}. If \code{TRUE} then return a Forest Plot of the estimates and confidence intervals; if \code{FALSE} return a table (default).}

\item{raw}{\code{logical}. If \code{TRUE} then return the values in a \code{data.frame} object, if \code{FALSE} return the values in a \code{gt} table object (default).}

\item{ci}{The confidence interval level required.}

\item{ci_normal}{\code{logical}. If \code{TRUE} then calculate the confidence interval based on asymptotic normality; if \code{FALSE} calculate the profile likelihood confidence interval using \code{confint} (default).}

\item{sigfig}{The number of significant figures to round the results to.}

\item{expand}{\code{logical}. If \code{TRUE} then additionally return the standard errors, z values and p-values from the glm summary object.}
}
\description{
Function used to present the model coefficient estimates and confidence intervals of one or more \code{glm} objects in a gt table, data.frame or Forest Plot.
}
\examples{
viewCoef(fittedModel1, exp=T)
modelList <- list(fittedModel1, fittedModel2)
viewCoef(modelList, plot=T)

}
\keyword{coefficients,}
\keyword{confidence}
\keyword{forest}
\keyword{interval,}
\keyword{plot}
