% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCols.R
\name{addCols}
\alias{addCols}
\title{Add Columns to glm Data}
\usage{
addCols(fittedModel)
}
\arguments{
\item{fittedModel}{A fitted \code{glm} model object of family \code{binomial}.}
}
\description{
S3 Method for glm object used to copy the model dataframe and add additional columns such as the predicted probability and logit of the response.
Used by \code{plotPredictors} during model diagnostics.
}
\examples{
new_data <- addCols(fittedModel)

}
\keyword{fitted}
