% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.listGLM.R
\name{compare}
\alias{compare}
\title{Compare GLM models using various criteria}
\usage{
compare(fittedModels, expand = F, sigfig = 4)
}
\arguments{
\item{fittedModels}{An object of class 'listGLM' i.e. a list of fitted \code{glm} model object(s) of family \code{binomial}.}

\item{expand}{\code{logical}. If  \code{FALSE} (default), return the AIC, BIC and McFadden's Pseudo R2-values. If TRUE` then return additional Pseudo-R2 values.}

\item{sigfig}{The number of significant figures to round the results to.}
}
\description{
S3 method for class 'listGLM'. Get AIC, BIC and Pseudo-R-squared values for each model in 'listGLM' object.
}
\examples{
compare(listGLM(fittedModel1, fittedModel2, fittedModel3),expand=T)

}
\keyword{R-squared}
\keyword{aic,}
\keyword{bic,}
\keyword{compare,}
