% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGLM.R
\name{listGLM}
\alias{listGLM}
\title{List GLM models}
\usage{
listGLM(...)
}
\arguments{
\item{...}{one or more fitted \code{glm} model objects of family \code{binomial}.}
}
\description{
Function used to create list of glm objects of class 'listGLM'
}
\examples{
fit_glm1 <- glm(am ~ cyl + hp + wt, data=mtcars, family = binomial(link = "logit"))
fit_glm2 <- glm(am ~ cyl + hp, data=mtcars, family = binomial(link = "logit"))
x <- listGLM(fit_glm1, fit_glm2)


}
\keyword{glm}
\keyword{list}
