% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredictors.R
\name{plotPredictors}
\alias{plotPredictors}
\title{Function used to return plots of the logit of the response vs numerical predictors}
\usage{
plotPredictors(fittedModel, interactive = FALSE)
}
\arguments{
\item{fittedModel;}{a logistic regression model fitted using the glm function}

\item{interactive;}{a flag used to indicate whether and interactive plot is returned (TRUE) or standard plots (FALSE)}
}
\description{
Used to check assumptions of logistic regression are not violated
}
\examples{
my_plots <- plotPredictors(fittedModel)
my_plots[[1]]

}
\keyword{assumptions,}
\keyword{diagnostics,}
\keyword{linearity,}
\keyword{logit,}
\keyword{plots}
\keyword{predictor}
