% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredictors.R
\name{plotPredictors}
\alias{plotPredictors}
\title{Plot Logit of Response against Numerical Predictors}
\usage{
plotPredictors(fittedModel, smooth = T, interactive = F)
}
\arguments{
\item{fittedModel}{A fitted \code{glm} model object of family \code{binomial}.}

\item{smooth}{\code{logical}. If \code{TRUE} then include a loess smoothing line (default); if \code{FALSE} do not include the loess line.}

\item{interactive}{\code{logical}. If \code{TRUE} then return an interactive plot; if \code{FALSE}, return ggplot objects (default).}
}
\description{
\code{plotPredictors} is a generic function used to plot a model's predictors against the the response.
Current implementations include an S3 Method for glm objects, used to return plots of the logit of the response vs numerical predictors in the fitted model.
}
\examples{
fit_glm1 <- glm(am ~ cyl + hp + wt, data=mtcars, family = binomial(link = "logit"))
my_plots <- plotPredictors(fit_glm1)
my_plots[[1]]

}
\keyword{assumptions,}
\keyword{diagnostics,}
\keyword{linearity,}
\keyword{logit,}
\keyword{plots}
\keyword{predictor}
