% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredictors.R
\name{plotPredictors}
\alias{plotPredictors}
\title{Plot Logit against Numerical Predictors}
\usage{
plotPredictors(fittedModel, interactive = F)
}
\arguments{
\item{fittedModel}{A fitted \code{glm} model object of family \code{binomial}.}

\item{interactive}{\code{logical}. If \code{TRUE} then return an interactive plot; if \code{FALSE}, return ggplot objects (default).}
}
\description{
S3 Method on glm object used to return plots of the logit of the response vs numerical predictors. Used during model diagnostics.
}
\examples{
my_plots <- plotPredictors(fittedModel)
my_plots[[1]]

}
\keyword{assumptions,}
\keyword{diagnostics,}
\keyword{linearity,}
\keyword{logit,}
\keyword{plots}
\keyword{predictor}
